# scripts/openapi.schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLM Guardrails Gateway",
    "description": "Safety middleware for LLM traffic and agent skills \u2014 gateway, policy, and leak scanning.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/analytics/dashboard": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_analytics_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsDashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/top-blocked-reasons": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Top Blocked Reasons",
        "description": "Most frequent fired rules among blocked requests, with the most recent occurrence of each.",
        "operationId": "top_blocked_reasons_analytics_top_blocked_reasons_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopBlockedReason"
                  },
                  "title": "Response Top Blocked Reasons Analytics Top Blocked Reasons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/users": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "User Breakdown",
        "description": "Per-user request + token breakdown for org admins.",
        "operationId": "user_breakdown_analytics_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/logs": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Logs",
        "description": "Paginated request log with status, keyword, date range, and backend filters.",
        "operationId": "logs_analytics_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "backend_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Backend Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/export": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Export Logs",
        "description": "Download request logs as a CSV file.",
        "operationId": "export_logs_analytics_export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/false-positive-candidates": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "False Positive Candidates",
        "description": "Blocked requests that users later disputed \u2014 either via positive feedback\n(thumbs-up on a blocked request) or because the owner later added the rule\nto their always-allow overrides. Grouped by fired rule so you can spot the\nguardrails most likely throwing false positives.",
        "operationId": "false_positive_candidates_analytics_false_positive_candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsDashboard": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/UsageSummary"
          },
          "time_series": {
            "items": {
              "$ref": "#/components/schemas/TimeSeriesPoint"
            },
            "type": "array",
            "title": "Time Series"
          },
          "top_rules": {
            "items": {
              "$ref": "#/components/schemas/TopFiredRule"
            },
            "type": "array",
            "title": "Top Rules"
          },
          "provider_usage": {
            "items": {
              "$ref": "#/components/schemas/ProviderUsage"
            },
            "type": "array",
            "title": "Provider Usage",
            "default": []
          },
          "recent_suspicious": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Suspicious",
            "default": []
          },
          "recent_logs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Logs"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "time_series",
          "top_rules",
          "recent_logs"
        ],
        "title": "AnalyticsDashboard"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderUsage": {
        "properties": {
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "tokens": {
            "type": "integer",
            "title": "Tokens"
          }
        },
        "type": "object",
        "required": [
          "backend",
          "model",
          "count",
          "tokens"
        ],
        "title": "ProviderUsage"
      },
...
      "TimeSeriesPoint": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "delivered": {
            "type": "integer",
            "title": "Delivered"
          },
          "blocked": {
            "type": "integer",
            "title": "Blocked"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "total",
          "delivered",
          "blocked"
        ],
        "title": "TimeSeriesPoint"
      },
      "TopBlockedReason": {
        "properties": {
          "fired_rule": {
            "type": "string",
            "title": "Fired Rule"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "last_occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Occurred At"
          }
        },
        "type": "object",
        "required": [
          "fired_rule",
          "count",
          "last_occurred_at"
        ],
        "title": "TopBlockedReason"
      },
      "TopFiredRule": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "count"
        ],
        "title": "TopFiredRule"
      },
...
      "UsageSummary": {
        "properties": {
          "total_requests": {
            "type": "integer",
            "title": "Total Requests"
          },
          "delivered": {
            "type": "integer",
            "title": "Delivered"
          },
          "input_blocked": {
            "type": "integer",
            "title": "Input Blocked"
          },
          "output_blocked": {
            "type": "integer",
            "title": "Output Blocked"
          },
          "rate_limited": {
            "type": "integer",
            "title": "Rate Limited",
            "default": 0
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          },
          "block_rate_pct": {
            "type": "number",
            "title": "Block Rate Pct"
          },
          "avg_latency_ms": {
            "type": "number",
            "title": "Avg Latency Ms"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "total_requests",
          "delivered",
          "input_blocked",
          "output_blocked",
          "error_count",
          "block_rate_pct",
          "avg_latency_ms",
          "total_tokens"
        ],
        "title": "UsageSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}